feat: Add Redis client for caching
This commit adds a new file  which contains the implementation of a Redis client for caching. The Redis client is used by the  utility module. It has dependencies on the  package.

The Redis client is implemented as a class  with methods for checking if the client is alive, getting a value from the cache, setting a value in the cache, and deleting a value from the cache.

A new instance of the  class is created and exported as  for use in other parts of the codebase.

This commit addresses the need for caching functionality in the application and improves performance by utilizing Redis as a caching solution.