chore: Refactor Redis client initialization
Refactor the Redis client initialization in the  file to use the  method from the  package. This change improves the readability and maintainability of the code.

The refactored code also includes error handling for the Redis client connection. If an error occurs, it will be logged to the console and the  flag will be set to . On successful connection, the  flag will be set to .

This commit addresses the need for a more robust Redis client initialization and improves the error handling for the Redis connection.